changes in documenatation
Additional Information: Some changes are required in this section regarding the data that is written in paragraph should be written in bullets and their are indentation issues with the pictures and form/field info.

Basic flow: Inside basic flow their is a link missing inside it if the enviornment is not given.

Navigation: Inside this page their are information issues picture is different and the info along with it is different and a additional voice control option is also added

Quick Start: Inside this page the step Verify your Email is missing and the subscription heading is the main heading not the point

Scanning: Inside this page for steps their are numbers used instead of bullets but in all the documentation we use Bullets.